fixing padding for the bottomsheet
