+++ android/app/src/main/java/org/uwaterloo/subletr/pages/home/HomePageView.kt
@@ -10,7 +10,6 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentWidth
@@ -54,8 +53,7 @@ fun HomePageView(
 ) {
 	val isListView = remember { mutableStateOf(true) }
 	Scaffold(
-		modifier = modifier
-			.imePadding(),
+		modifier = modifier,
 		topBar = {
 			Row(
 				modifier = Modifier
@@ -83,8 +81,7 @@ fun HomePageView(
 					viewModel = viewModel.homeListChildViewModel,
 					uiState = uiState,
 				)
-			}
-			else if (uiState is HomeMapUiState) {
+			} else if (uiState is HomeMapUiState) {
 				HomeMapChildView(
 					modifier = Modifier.padding(padding),
 					viewModel = viewModel.homeMapChildViewModel,
